PP-8229 refactor GatewayAccountCredentialsResource
- move GatewayAccountCredentialsResource to gatewayaccountcredentials
  package.
- move other relevant classes as well
- move existing tests to new package
