+++ src/main/java/uk/gov/pay/connector/gatewayaccountcredentials/resource/GatewayAccountCredentialsRequestValidator.java
@@ -1,4 +1,4 @@
-package uk.gov.pay.connector.gatewayaccount.resource;
+package uk.gov.pay.connector.gatewayaccountcredentials.resource;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.google.common.base.Joiner;
